+++ packages/address-extractor/tests/extractor/utils.mock.ts
@@ -1,4 +1,5 @@
 import { DataSource } from '@rosen-bridge/extended-typeorm';
+
 import { BoxEntity, migrations } from '../../lib';
 
 export const createDatabase = async (): Promise<DataSource> => {