+++ packages/observation-extractors/evm-observation-extractor/tests/testData.ts
@@ -41,7 +41,7 @@ export const rosenData = {
   targetChainTokenId: 'target-token-id',
   sourceTxId:
     '0x3b194eea7cf9507e745806265738ca19213be209885534161ec0fa9c232c9fea',
-  rawData: 'raw-data',
+  rawData: 'mocked',
 };
 
 export const expectedObservation = {
@@ -61,4 +61,5 @@ export const expectedObservation = {
   block: 'block-id',
   requestId: Buffer.from(blake2b(tx.hash!, undefined, 32)).toString('hex'),
   extractor: 'test-observation-extractor',
+  rawData: 'mocked',
 };