+++ packages/scanners/cardano-scanner/lib/scanner/blockfrost.ts
@@ -1,5 +1,6 @@
-import { BlockFrostTransaction } from '../interfaces/BlockFrost';
 import { GeneralScanner, ScannerConfig } from '@rosen-bridge/abstract-scanner';
+
+import { BlockFrostTransaction } from '../interfaces/blockFrost';
 
 class CardanoBlockFrostScanner extends GeneralScanner<BlockFrostTransaction> {
   constructor(config: ScannerConfig<BlockFrostTransaction>) {