+++ packages/scanners/evm-scanner/tests/testRpcNetwork.ts
@@ -1,0 +1,5 @@
+import { EvmRpcNetwork } from '../lib/evmRpcNetwork';
+
+export class TestEvmRpcNetwork extends EvmRpcNetwork {
+  getProvider = () => this.provider;
+}