+++ packages/observation-extractors/bitcoin-runes-observation-extractor/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @rosen-bridge/bitcoin-runes-observation-extractor
 
+## 0.3.0
+
+### Minor Changes
+
+- Add rawData to the return data of the extractor's `processTransactions` method
+
+### Patch Changes
+
+- Fix package-lock and move typescript and types/node into root
+- Update eslint and plugins:
+  - Apply new rules such as sort imports and file name
+- Update dependencies
+  - @rosen-bridge/extended-typeorm@1.0.1
+  - @rosen-clients/rate-limited-axios@1.1.0
+  - @rosen-bridge/rosen-extractor@10.1.1
+  - @rosen-bridge/tokens@4.0.1
+  - @rosen-bridge/abstract-logger@3.0.1
+  - @rosen-bridge/json-bigint@1.1.0
+  - @rosen-clients/ergo-explorer@2.1.0
+  - @rosen-clients/ergo-node@3.1.0
+  - @rosen-bridge/abstract-observation-extractor@0.2.0
+  - @rosen-bridge/bitcoin-scanner@0.2.1
+  - @rosen-bridge/scanner-interfaces@0.2.1
+
+## 0.2.0
+
+### Minor Changes
+
+- Use @rosen-clients/rate-limited-axios instead of @rosen-bridge/rate-limited-axios
+
+### Patch Changes
+
+- Handle Unisat indexed height (it throws error if the given block is not indexed by Unisat yet)
+- Change unisatApiKey header to Authorization and Bearer
+- Update dependencies
+  - @rosen-bridge/scanner-interfaces@0.2.0
+  - @rosen-bridge/abstract-observation-extractor@0.1.2
+  - @rosen-bridge/bitcoin-scanner@0.2.0
+  - @rosen-bridge/extended-typeorm@1.0.0
+  - @rosen-bridge/tokens@4.0.0
+  - @rosen-bridge/rosen-extractor@10.0.0
+  - @rosen-bridge/json-bigint@1.0.0
+  - @rosen-bridge/abstract-logger@3.0.0
+
 ## 0.1.1
 
 ### Patch Changes